requirements.txt
django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
qrcode==8.0
pillow==11.0.0
python-barcode==0.15.1
django-cors-headers==4.6.0
reportlab==4.2.5